calendar: be smarter about how conflicting events are displayed
in a set of conflicting events, we can now internally merge events which don't
overlap with each other

for example, say you have event A (from 3 pm to 5 pm), event B (from 3:30 pm to
4 pm), and event C (from 4:30 to 5 pm). before, we'd display events A, B, and C
in three separate columns, which wastes space. now, we are smart enough to
display event A in one column, and events B and C together in a second column.

this closes #163